修复生成class的py文件的问题

+++ RinaChanBoardHardware/RinaChanBoardHardwareV2/include/led.h
@@ -24,7 +24,7 @@ void init_led(CRGB leds[], CRGB color);
  * @param length    [in]
 
  */
-void decodeFaceHex(const char hexBytes[], int (&cells)[16][18], size_t length);
+void decodeFaceHex(const char hexBytes[], uint8_t (&cells)[16][18], size_t length);
 
 /**
  * @brief 通过RGB值直接更新led颜色
@@ -49,7 +49,7 @@ void updateColor(CRGB leds[], const uint8_t &R, const uint8_t &G, const uint8_t 
  * @param leds 目标led
  * @param color 目标颜色
  */
-void face_update(int face[16][18], CRGB leds[], CRGB color);
+void face_update(uint8_t face[16][18], CRGB leds[], CRGB color);
 
 /**
  * @brief 从状态字符串更新璃奈板表情
